:triangular_ruler: ModCache: Removed unused field 'materials'
The mysterious function `Ogre::String CacheSystem::addMeshMaterials(CacheEntry& entry, Ogre::Entity* e)` which used it was never called.

Also fixed compiler warning in Utils.cpp

+++ source/main/utils/Utils.cpp
@@ -59,7 +59,7 @@ String HashFile(const char *szFileName)
 
     if (ulFileSize != 0)
     {
-        ulBlocks = ulFileSize / uData.size();
+        ulBlocks = ulFileSize / static_cast<unsigned long>(uData.size());
         ulRest = ulFileSize % uData.size();
     }
     else
